Added callback function to timer

+++ src/main/java/unboxthecat/meowoflegends/GameState.java
@@ -10,15 +10,15 @@ import java.util.Map;
 import java.util.UUID;
 
 public class GameState implements Listener {
-    private static final int SERVER_TICK_RATE = 20; //may need to dynamically capture it in the future
+    private static final long SERVER_TICK_RATE = 20; //may need to dynamically capture it in the future
     private static final Map<UUID, MOLEntity> players = new HashMap<>();
     private GameState() {};
 
-    public static int secondToTick(double seconds) {
-        return (int)(Math.round(seconds * (double)SERVER_TICK_RATE));
+    public static long secondToTick(double seconds) {
+        return Math.round(seconds * (double)SERVER_TICK_RATE);
     }
 
-    public static double tickToSecond(int tick) {
+    public static double tickToSecond(long tick) {
         return (double)tick / 20.0;
     }
