+++ src/main/java/unboxthecat/meowoflegends/handler/LoginMessageHandler.java
@@ -16,14 +16,13 @@ public class LoginMessageHandler implements Listener {
             Bukkit.getOnlinePlayers().forEach(player ->
                 player.sendTitle("Welcome",
                     event.getPlayer().getName(),
-                    GameState.secondToTick(1.0),
-                    GameState.secondToTick(5.0),
-                    GameState.secondToTick(1.0))
+                        (int) GameState.secondToTick(1.0),
+                        (int) GameState.secondToTick(5.0),
+                        (int) GameState.secondToTick(1.0))
             );
         }
         event.setJoinMessage(ChatColor.AQUA + event.getPlayer().getName() + ChatColor.GREEN + " hopped on the server!");
     }
-
     @EventHandler
     public void onPlayerQuit(PlayerQuitEvent event) {
         event.setQuitMessage(ChatColor.AQUA + event.getPlayer().getName() + ChatColor.RED + " rage quited!");
